file_organizer/pyproject.toml
[project]
name = "file-organizer"
version = "0.1.0"
description = "CLI tool that organizes files into structured directories by type"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
    "tqdm>=4.67.1",
]

[project.scripts]
file-organizer = "file_organizer.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/file_organizer"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-tmp-files>=0.0.2",
]
